% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admm_algorithm.R
\name{ADMM_algorithm}
\alias{ADMM_algorithm}
\title{ADMM algorithm for graphical SLOPE}
\usage{
ADMM_algorithm(sample_cov, lambda, mu = 1.1, max_iter = 10000, epsilon = 1e-04)
}
\arguments{
\item{sample_cov}{variance-covariance matrix.}

\item{lambda}{vector of SLOPE regularizers.}

\item{mu}{correction for lambda scaling.}

\item{max_iter}{maximum number of iterations of ADMM algorithm. Default 10 000.}

\item{epsilon}{a value determining stop condition.}
}
\value{
\code{ADMM_algorithm} returns a list with components:
\itemize{
\item \code{precision_matrix} precision matrix revealing graph structure for the data.
\item \code{iterations} number of iterations before stop condition or \code{max_iter}.
}
}
\description{
Executes ADMM algorithm to find precision matrix.
}
\details{
ADMM algorithm is used to solve convex optimization problems by dividing them into smaller, easier to solve problems.
}
\examples{
sample_cov = cov(scale(mtcars))
lambda = gslope::create_lambda(sample_cov, nrow(mtcars))
precision_matrix = ADMM_algorithm(sample_cov, lambda)
}
\keyword{admm}
