% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef}
\alias{coef}
\alias{coef.gslope}
\title{Prints precision matrix from gslope}
\usage{
coef(x)

\method{coef}{gslope}(x)
}
\arguments{
\item{x}{an object of class `'gslope'`}
}
\value{
Prints output on the screen
}
\description{
Prints precision matrix from gslope
}
\examples{
w = gslope(as.matrix(mtcars))
coef(w)

}
