% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lambda.R
\name{create_lambda}
\alias{create_lambda}
\title{Penalty parameters for graphical SLOPE}
\usage{
create_lambda(sample_cov, n, alpha = 0.05)
}
\arguments{
\item{sample_cov}{a sample variance-covariance matrix.}

\item{n}{number of observations.}

\item{alpha}{significance level.}
}
\value{
\code{create_lambda()} returns a vector of the length \eqn{p * (p-1)/2}, which is equal
to number of elements in lower (upper) triangle of a variance-covariance matrix excluding diagonal.
}
\description{
Computes penalty parameters \eqn{\lambda} for graphical SLOPE.
}
\details{
The Lambda series are computed based on Benjamini-Hochberg's procedure.
}
\examples{
sample_cov = cov(scale(mtcars))
create_lambda(sample_cov, nrow(mtcars))
}
\keyword{graphical}
\keyword{lambda,}
\keyword{penalty}
\keyword{slope,}
