% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gslope.R
\name{gslope}
\alias{gslope}
\title{Graphical SLOPE}
\usage{
gslope(
  data,
  lambda = NULL,
  sample_cov = cov(data),
  scaled = FALSE,
  mu = 1.1,
  max_iter = 10000,
  epsilon = 1e-04,
  threshold = 1e-04,
  alpha = 0.05
)
}
\arguments{
\item{data}{a matrix containing observations of variables of interest.}

\item{lambda}{vector of regularizers for SLOPE. By default computed based on Benjamini-Hochberg's method.}

\item{sample_cov}{variance-covariance matrix.}

\item{scaled}{logical. The data need to be scaled so that it has mean = 0 and variance = 1. If TRUE, build-in data scaling will be omitted.}

\item{mu}{## TODO:  czo to?}

\item{max_iter}{maximum number of iterations allowed in ADMM algorithm. Default 10 000.}

\item{epsilon}{a value used to determine accuracy of the ADMM algorithm. Default 10e-4.}

\item{threshold}{a value used to determine which variables in graphical model are not linked by an edge. The larger the value, the more variables will be unlinked. Default 10e-4.}

\item{alpha}{significance level.}
}
\value{
\code{gslope} returns a list containing following components:
\item{precision_matrix}{a precision matrix revealing graph structure for the data.}
\item{covariance_matrix}{covariance matrix equal to the inverse of the presicion matrix.}
\item{scaled_precision_matrix}{...}
\item{lambda}{a vector of penalty parameters used in SLOPE.}
\item{iterations}{a number of iterations performed in ADMM algorithm.}
}
\description{
Computes precision matrix for graph models using graphical SLOPE.
}
\details{
\code{gslope} selects high probability graph structure for graphical model with likelihood-based methods combined with ordered L1-regularization. Namely, it solves - using ADMM algorithm - the following  maximization problem:
\deqn{ log det \Theta - tr(S \Theta) - \lambda(\Theta), subject to \Theta \in S+,}
where S is a sample covariance matrix, \eqn{\lambda(\Theta)} is a series of regularizers for SLOPE and S+ denotes a set of symmetric, semidefinite matrices.
}
\examples{
gslope(mtcars, epsilon = 1e-3)
}
\keyword{graphical}
\keyword{matrix,}
\keyword{precision}
\keyword{slope}
