% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gslope_admm.R
\name{gslope_admm}
\alias{gslope_admm}
\title{Graphical SLOPE}
\usage{
gslope_admm(C, Z, U, lambda_seq, rho, max_iter = 100, tol_infeas = 0.001,
  verbose = FALSE)
}
\arguments{
\item{C}{sample covariance matrix}

\item{Z}{initial estimate of inverse covariance}

\item{U}{initial estimate of covariance}

\item{lambda_seq}{sequence of lambda values of ordered L1 norm}

\item{rho}{constant}

\item{max_iter}{maximum number of ADMM iterations}

\item{tol_infeas}{tolerance to feasibility rules}

\item{verbose}{should additional info be printed out}
}
\description{
Method for estimating sparse inverse covariance matrix
}
\details{
Implementation using alternating direction method of multipliers
}
