% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\title{Plot precision matrix}
\usage{
plot(x, col)
}
\arguments{
\item{x}{an object of class `'gslope'`}

\item{col}{a character, plot color. Default "black".}
}
\value{
Prints output on the screen
}
\description{
Plot precision matrix
}
\examples{
bla bla
}
