% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gslope.R
\name{prepare_lambda}
\alias{prepare_lambda}
\title{Preparation of lambda}
\usage{
prepare_lambda(lambda, low_tri_size)
}
\arguments{
\item{lambda}{a vector of SLOPE regularizers.}

\item{low_tri_size}{a number of elements in lower (upper) triangle of a variance-covariance matrix excluding diagonal.\cr
Should be equal to \eqn{p * (p-1)/2}, where p in a number of variables (columns) in the data.}
}
\value{
\code{prepare_lambda} returns a sorted descending vector of the length \eqn{p * (p-1)/2}.
}
\description{
Prepares penalty parameters \eqn{\lambda} for further computations in graphical SLOPE.
}
\details{
In case of vector lambda being too short, it is filled with zeros up to the set length. If the vector is too long, it is cut to the proper size.
}
\examples{
v <- sample(1:100,10)
prepare_lambda(v,10)
prepare_lambda(v,5)
prepare_lambda(v,17)
}
\keyword{lambda}
