% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print}
\alias{print}
\alias{print.gslope}
\title{Prints summary from gslope}
\usage{
print(x)

\method{print}{gslope}(x)
}
\arguments{
\item{x}{an object of class `'gslope'`}
}
\value{
Prints output on the screen
}
\description{
Prints summary from gslope
}
\examples{
bla bla

}
