% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope_admm.R
\name{slope_admm}
\alias{slope_admm}
\title{Graphical SLOPE}
\usage{
slope_admm(A, b, z, u, lambda_seq, rho, max_iter = 100, tol_infeas = 0.001,
  verbose = FALSE)
}
\arguments{
\item{A}{design matrix}

\item{b}{response}

\item{z}{initial value}

\item{u}{initial value}

\item{lambda_seq}{sequence of lambda values for sorted L1 norm}

\item{rho}{constant}

\item{max_iter}{maximum number of ADMM iterations}

\item{tol_infeas}{tolerance to feasibility rules}

\item{verbose}{should additional info be printed out}
}
\description{
Method for coefficients in regression model
}
\details{
Implementation using alternating direction method of multipliers
}
