% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary}
\alias{summary}
\alias{summary.gslope}
\title{Prints all parameters from gslope}
\usage{
summary(x)

\method{summary}{gslope}(x)
}
\arguments{
\item{x}{an object of class `'gslope'`}
}
\value{
Prints following output on the screen:
\itemize{

\item \code{call} formula used in function call.
\item \code{precision matrix} a sparse precision matrix
\item \code{covariance matrix} covariance matrix for the data
\item \code{scaled precision matrix} see: ?gslope
\item \code{lambda} lambda regularizers used
\item \code{iterations} number of iterations
\item \code{graph} a graph object from igraph package. For more details see ?igraph.
\item \code{clusters} clusters revealing sub-graph structure in the model For more details see ?gslope.
}
}
\description{
Prints all parameters from gslope
}
\examples{
d = gslope(as.matrix(mtcars))
summary(d)

}
