% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef}
\alias{coef}
\title{Prints precision matrix from gslope}
\usage{
coef(x)
}
\arguments{
\item{x}{an object of class `'gslope'`}
}
\value{
Prints output on the screen
}
\description{
Prints precision matrix from gslope
}
\examples{
bla bla

}
