% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_plot.R
\name{graph_plot}
\alias{graph_plot}
\title{Plot graph for gslope.}
\usage{
graph_plot(x)

graph_plot(x)
}
\arguments{
\item{x}{an object of class `'gslope'`}
}
\value{
Draw graph.
}
\description{
Plot graph for gslope.
}
\examples{
w = gslope(as.matrix(mtcars))
graph_plot(w)
}
