% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gslope-package.R
\docType{package}
\name{gslope}
\alias{gslope}
\alias{gslope-package}
\title{Gaussian graphical model estimation using sorted l1 penalty}
\description{
Gaussian graphical model estimation using sorted l1 penalty
}
\details{
Version: 0.2.0
}
\author{
Piotr Sobczyk, Malgorzata Bogdan

  Maintainer: Piotr Sobczyk \email{Piotr.Sobczyk@pwr.edu.pl}
}
