% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary}
\alias{summary}
\alias{summary.gslope}
\title{Prints all parameters from gslope}
\usage{
summary(x)

\method{summary}{gslope}(x)
}
\arguments{
\item{x}{an object of class `'gslope'`}
}
\value{
Prints output on the screen
}
\description{
Prints all parameters from gslope
}
\examples{
d=gslope(as.matrix(mtcars))
summary(d)

}
